+++ mitmproxy/addons/clientplayback.py
@@ -133,7 +133,10 @@ class ClientPlayback:
             self.inflight = await self.queue.get()
             try:
                 h = ReplayHandler(self.inflight, self.options)
-                await h.replay()
+                if ctx.options.client_replay_concurrency == -1:
+                    asyncio_utils.create_task(h.replay(), name="client playback awaiting response")
+                else:
+                    await h.replay()
             except Exception:
                 ctx.log(f"Client replay has crashed!\n{traceback.format_exc()}", "error")
             self.queue.task_done()
@@ -160,6 +163,10 @@ class ClientPlayback:
             "client_replay", typing.Sequence[str], [],
             "Replay client requests from a saved file."
         )
+        loader.add_option(
+            "client_replay_concurrency", int, 1,
+            "Concurrency limit on in-flight client replay requests. Currently the only valid values are 1 and -1 (no limit)."
+        )
 
     def configure(self, updated):
         if "client_replay" in updated and ctx.options.client_replay:
@@ -168,6 +175,10 @@ class ClientPlayback:
             except exceptions.FlowReadException as e:
                 raise exceptions.OptionsError(str(e))
             self.start_replay(flows)
+
+        if "client_replay_concurrency" in updated:
+            if ctx.options.client_replay_concurrency not in [-1, 1]:
+                raise exceptions.OptionsError("Currently the only valid client_replay_concurrency values are -1 and 1.")
 
     @command.command("replay.client.count")
     def count(self) -> int: